GRPC_emulated/GRPC-l2-Server/requirements.txt
numpy 
numba
grpcio
grpcio-tools
motor
pymongo
python-dotenv
